Module task_maker::executor

Structs

ChannelFileIterator
An iterator over the byte chunks sent during the FileProtocol mode in a channel

ChannelFileSender
Utility function to send a file to a channel using FileProtocol

Executor
The Executor is the main component of the server, this will receive the DAG to evaluate and will schedule the tasks to the workers, sending to the client the responses

ExecutorClient
This is a client of the Executor, the client is who sends a DAG for an evaluation, provides some files and receives the callbacks from the server. When the server notifies a callback function is called by the client.

ExecutorData
Internal state of the Executor

ExecutorStatus
The current status of the Executor, this is sent to the user when the status is asked

LocalExecutor
An Executor that runs locally

Sandbox
Wrapper around the sandbox. Cloning this struct will keep the reference of the same sandbox, keeping the content alive.

Worker
The worker is the component that receives the work from the server and sends the results back.

WorkerConn
An handle of the connection to the worker

WorkerJob
A job that is sent to a worker, this should include all the information the worker needs to start the evaluation

WorkerState
State of a worker as seen by the server

Enums

ExecutorClientMessage
Messages that the client sends to the server

ExecutorServerMessage
Messages that the server sends to the client

FileProtocol
Messages sent during the FileProtocol operation, ie during the transfer of a file

SandboxResult
Result of the execution of the sandbox

WorkerClientMessage
Messages sent by the workers to the server

WorkerError
An error generated by the worker

WorkerServerMessage
Messages sent by the server to the worker

Functions

deserialize_from
Deserialize a message from the channel and return it.

serialize_into
Serialize a message into the sender serializing it.

strsignal
Returns a string with the text representation of the signal.

Type Definitions

ChannelReceiver
The channel part that receives data.

ChannelSender
The channel part that sends data.