Struct task_maker::opt::Opt

pub struct Opt {
    pub task_dir: PathBuf,
    pub ui: UIType,
    pub keep_sandboxes: bool,
    pub dry_run: bool,
    pub no_cache: Option<Option<String>>,
    pub exclusive: bool,
    pub extra_time: Option<f64>,
    pub copy_exe: bool,
    pub copy_logs: bool,
    pub filter: Vec<String>,
    pub solution: Vec<PathBuf>,
    pub max_depth: u32,
    pub clean: bool,
    pub store_dir: Option<PathBuf>,
    pub max_cache: u64,
    pub min_cache: u64,
    pub num_cores: Option<usize>,
    pub booklet_solutions: bool,
    pub no_statement: bool,
    pub verbose: u8,
    pub evaluate_on: Option<String>,
    pub name: Option<String>,
    pub remote: Option<Remote>,
    pub task_info: bool,
    pub sandbox: bool,
}

Fields
task_dir: PathBuf

Directory of the task to evaluate

ui: UIType

Which UI to use, available UIS are: print, raw, curses, json.

Note that the JSON api is not stable yet.

keep_sandboxes: bool

Keep all the sandbox directories

dry_run: bool

Do not write any file inside the task directory

no_cache: Option<Option<String>>

Disable the cache for this comma separated list of tags

Providing an empty list will disable all the caches. The supported tags are: compilation, generation, evaluation, checking, booklet.

exclusive: bool

Do not run in parallel time critical executions on the same machine

extra_time: Option<f64>

Give to the solution some extra time before being killed

copy_exe: bool

Copy the executables to the bin/ folder

copy_logs: bool

Copy the logs of some executions to the bin/logs/ folder

filter: Vec<String>

Execute only the solutions whose names start with the filter

Note that just the file name is checked (e.g. sol.cpp is the same as sol/sol.cpp). Without specifying anything all the solutions are executed.

solution: Vec<PathBuf>

Evaluate only the solution with the specified path

The solution can reside anywhere in the filesystem.

max_depth: u32

Look at most for this number of parents for searching the task

clean: bool

Clear the task directory and exit

store_dir: Option<PathBuf>

Where to store the storage files, including the cache

max_cache: u64

Maximum size of the storage directory, in MiB

min_cache: u64

When the storage is flushed, this is the new maximum size, in MiB.

num_cores: Option<usize>

The number of CPU cores to use

booklet_solutions: bool

Include the solutions in the booklet

no_statement: bool

Do not build the statement files and the booklets

verbose: u8

Verbose mode (-v, -vv, -vvv, etc.). Note that it does not play well with curses ui.

evaluate_on: Option<String>

Run the evaluation on a remote server instead of locally

name: Option<String>

The name to use for the workers or for the client in remote executions

This option only has no effect with --server.

remote: Option<Remote>

Options for spawning a client or a server.

task_info: bool

Show task information

sandbox: bool

Run the sandbox instead of the normal task-maker.

This option is left as undocumented as it's not part of the public API.

Methods

impl Opt

pub fn to_config(&self) -> EvaluationConfig

Make an EvaluationConfig from this command line options.

pub fn store_dir(&self) -> PathBuf

Get the store directory of this configuration. If nothing is specified a cache directory is used if available, otherwise a temporary directory.

pub fn enable_log(&mut self)