Trait task_maker::score_types::ScoreType

pub trait ScoreType<SubtaskId: Eq + PartialOrd + Hash + Copy, TestcaseId: Eq + PartialOrd + Hash + Copy>: Debug {
    fn testcase_score(
        &mut self, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        score: f64
    );

    fn get_subtask_score(&mut self, callback: Box<dyn Fn(SubtaskId, f64)>);

    fn get_task_score(&mut self, callback: Box<dyn Fn(f64)>);

    fn boxed(&self) -> Box<dyn ScoreType<SubtaskId, TestcaseId>>;
}

A score type is the aggregation function for the testcases scores. From the testcases score this is able to compute the subtask scores and the task total score.

Required methods

fn testcase_score(
&mut self,
subtask: SubtaskId,
testcase: TestcaseId,
score: f64
)

Tell the ScoreType the score of a new testcase. Will return true if the score of a subtask has become ready.

fn get_subtask_score(&mut self, callback: Box<dyn Fn(SubtaskId, f64)>)

Set the callback that will be called when a subtask has the score ready.

fn get_task_score(&mut self, callback: Box<dyn Fn(f64)>)

Set the callback that will be called when the task has the score ready.

fn boxed(&self) -> Box<dyn ScoreType<SubtaskId, TestcaseId>>

Clone this ScoreType inside a Box. Note that the callbacks will be resetted in the result.

Loading content...