Trait task_maker::task_types::Checker

pub trait Checker<SubtaskId, TestcaseId>: Debug where
    SubtaskId: Eq + PartialOrd + Hash + Copy,
    TestcaseId: Eq + PartialOrd + Hash + Copy, {
    fn check(
        &self, 
        eval: &mut EvaluationData, 
        input: File, 
        output: Option<File>, 
        test: File, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        callback: BoxFnOnce<'static, (CheckerResult,)>
    ) -> Execution;
}

A trait that describes what is a checker: something that given an input file, an optional correct output file and the contestant's output file computes a score (and eventually message) for that testcase.

Required methods

fn check(
&self,
eval: &mut EvaluationData,
input: File,
output: Option<File>,
test: File,
subtask: SubtaskId,
testcase: TestcaseId,
callback: BoxFnOnce<'static, (CheckerResult,)>
) -> Execution

Add the checking process to the DAG and call the callback when the checker is done

Loading content...