Trait task_maker::task_types::Task

pub trait Task<SubtaskId: Eq + PartialOrd + Hash + Copy + Debug + 'static, TestcaseId: Eq + PartialOrd + Hash + Copy + Debug + 'static, SubtaskInfo: SubtaskInfo, TestcaseInfo: TestcaseInfo<SubtaskId, TestcaseId>>: Debug {
    fn format() -> &'static str
    where
        Self: Sized;

    fn path(&self) -> &Path;

    fn name(&self) -> String;

    fn title(&self) -> String;

    fn subtasks(&self) -> &HashMap<SubtaskId, SubtaskInfo>;

    fn testcases(
        &self, 
        subtask: SubtaskId
    ) -> &HashMap<TestcaseId, TestcaseInfo>;

    fn score_type(&self) -> &dyn ScoreType<SubtaskId, TestcaseId>;

    fn solutions(
        &self
    ) -> &HashMap<PathBuf, Box<dyn Solution<SubtaskId, TestcaseId>>>;

    fn checker(
        &self, 
        subtask: SubtaskId, 
        testcase: TestcaseId
    ) -> &Box<dyn Checker<SubtaskId, TestcaseId>>;

    fn get_ui_interface(
        &self
    ) -> Arc<dyn TaskUIInterface<SubtaskId, TestcaseId>>;

    fn evaluate(
        &self, 
        eval: EvaluationData, 
        options: &dyn EvaluationOptions, 
        executor: LocalExecutor
    ) { ... }
}

Trait that describes a generic task. Every task must have a generator (a way of getting testcases) and can have a validator, an official solution, but has to have a checker that assigns a score to a solution.

Required methods

fn format() -> &'static str where
Self: Sized,

Name of the format of the task

fn path(&self) -> &Path

Path to the root folder of the task.

fn name(&self) -> String

Name of the task (the short one)

fn title(&self) -> String

Title of the task (the long one)

fn subtasks(&self) -> &HashMap<SubtaskId, SubtaskInfo>

The list of the subtasks for this task

fn testcases(&self, subtask: SubtaskId) -> &HashMap<TestcaseId, TestcaseInfo>

The list of the testcases for that subtask

fn score_type(&self) -> &dyn ScoreType<SubtaskId, TestcaseId>

The score type to use for this task.

fn solutions(
&self
) -> &HashMap<PathBuf, Box<dyn Solution<SubtaskId, TestcaseId>>>

The list of known solution files

fn checker(
&self,
subtask: SubtaskId,
testcase: TestcaseId
) -> &Box<dyn Checker<SubtaskId, TestcaseId>>

The optional checker that will check the output file

fn get_ui_interface(&self) -> Arc<dyn TaskUIInterface<SubtaskId, TestcaseId>>

Get the TaskUIInterface relative to this

Loading content...
Provided methods

fn evaluate(
&self,
eval: EvaluationData,
options: &dyn EvaluationOptions,
executor: LocalExecutor
)

Build the DAG of the evaluation of this task and use the executor to start the evaluation. This method will block until the evaluation ends.

Loading content...