Trait task_maker::task_types::TaskFormat

pub trait TaskFormat {
    type SubtaskId: Eq + PartialOrd + Hash + Copy;
    type TestcaseId: Eq + PartialOrd + Hash + Copy;
    type SubtaskInfo: SubtaskInfo;
    type TestcaseInfo: TestcaseInfo<Self::SubtaskId, Self::TestcaseId>;
    fn is_valid(path: &Path) -> bool;

    fn parse(
        path: &Path
    ) -> Result<Box<dyn Task<Self::SubtaskId, Self::TestcaseId, Self::SubtaskInfo, Self::TestcaseInfo>>, Error>;
}

A task format is a way of laying files in a task folder, every task folder contains a single task which type can be different even for the same format. For example in a IOI-like format there could be a Batch task, a Communication task, ...

Associated Types

type SubtaskId: Eq + PartialOrd + Hash + Copy

Type of the identifier of a subtask

type TestcaseId: Eq + PartialOrd + Hash + Copy

Type of the identifier of a testcase

type SubtaskInfo: SubtaskInfo

Type of the information about a subtask

type TestcaseInfo: TestcaseInfo<Self::SubtaskId, Self::TestcaseId>

Type of the information about a testcase

Loading content...
Required methods

fn is_valid(path: &Path) -> bool

Whether the path points to a valid task for this format.

fn parse(
path: &Path
) -> Result<Box<dyn Task<Self::SubtaskId, Self::TestcaseId, Self::SubtaskInfo, Self::TestcaseInfo>>, Error>

Assuming path is valid make a Task from that directory.

Loading content...