Trait task_maker::task_types::TaskUIInterface

pub trait TaskUIInterface<SubtaskId: Eq + PartialOrd + Hash + Copy + Debug + 'static, TestcaseId: Eq + PartialOrd + Hash + Copy + Debug + 'static> {
    fn task_info(&self, sender: Arc<Mutex<UIMessageSender>>);

    fn generation_result(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        status: UIExecutionStatus
    );

    fn validation_result(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        status: UIExecutionStatus
    );

    fn solution_result(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        status: UIExecutionStatus
    );

    fn evaluation_result(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        solution: PathBuf, 
        status: UIExecutionStatus
    );

    fn checker_result(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        solution: PathBuf, 
        status: UIExecutionStatus
    );

    fn testcase_score(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        testcase: TestcaseId, 
        solution: PathBuf, 
        score: f64
    );

    fn subtask_score(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        subtask: SubtaskId, 
        solution: PathBuf, 
        score: f64
    );

    fn task_score(
        &self, 
        sender: Arc<Mutex<UIMessageSender>>, 
        solution: PathBuf, 
        score: f64
    );
}

An interface between a Task type and the UI. Each task type may send different information to the UI based on the number of the testcase.

Required methods

fn task_info(&self, sender: Arc<Mutex<UIMessageSender>>)

Send the information about the task to the UI.

fn generation_result(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
status: UIExecutionStatus
)

Send to the UI the status of the generation of a testcase.

fn validation_result(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
status: UIExecutionStatus
)

Send to the UI the status of the validation of a testcase.

fn solution_result(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
status: UIExecutionStatus
)

Send to the UI the status of the generation of the output of a testcase.

fn evaluation_result(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
solution: PathBuf,
status: UIExecutionStatus
)

Send to the ui the status of the evaluation of a solution on a testcase.

fn checker_result(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
solution: PathBuf,
status: UIExecutionStatus
)

Send to the ui the status of the checking of a solution on a testcase.

fn testcase_score(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
testcase: TestcaseId,
solution: PathBuf,
score: f64
)

Send to the ui the score of a solution of a testcase.

fn subtask_score(
&self,
sender: Arc<Mutex<UIMessageSender>>,
subtask: SubtaskId,
solution: PathBuf,
score: f64
)

Send to the ui the score of a solution of a subtask.

fn task_score(
&self,
sender: Arc<Mutex<UIMessageSender>>,
solution: PathBuf,
score: f64
)

Send to the ui the score of a solution of a task.

Loading content...