Trait task_maker::task_types::TestcaseInfo

pub trait TestcaseInfo<SubtaskId: Eq + PartialOrd + Hash + Copy, TestcaseId: Eq + PartialOrd + Hash + Copy> {
    fn write_input_to(&self) -> Option<PathBuf>;

    fn write_output_to(&self) -> Option<PathBuf>;

    fn generator(&self) -> Arc<dyn Generator<SubtaskId, TestcaseId>>;

    fn validator(&self) -> Option<Arc<dyn Validator<SubtaskId, TestcaseId>>>;

    fn solution(&self) -> Arc<dyn Solution<SubtaskId, TestcaseId>>;
}

Some basic information about a testcase.

Required methods

fn write_input_to(&self) -> Option<PathBuf>

Write the input file to this path if it's not a dry-run, relative to the task directory.

fn write_output_to(&self) -> Option<PathBuf>

Write the output file to this path if it's not a dry-run, relative to the task directory.

fn generator(&self) -> Arc<dyn Generator<SubtaskId, TestcaseId>>

The generator that will create that testcase

fn validator(&self) -> Option<Arc<dyn Validator<SubtaskId, TestcaseId>>>

The optional validator that will validate that testcase

fn solution(&self) -> Arc<dyn Solution<SubtaskId, TestcaseId>>

The solution of the testcase.

Loading content...