Trait task_maker::task_types::Validator

pub trait Validator<SubtaskId, TestcaseId>: Debug where
    SubtaskId: Eq + PartialOrd + Hash + Copy,
    TestcaseId: Eq + PartialOrd + Hash + Copy, {
    fn validate(
        &self, 
        eval: &mut EvaluationData, 
        input: File, 
        subtask: SubtaskId, 
        testcase: TestcaseId
    ) -> (File, Option<Execution>);
}

A trait that describes what is a validator: something that known which testcase and given that input file checks if it respects all constraints.

Required methods

fn validate(
&self,
eval: &mut EvaluationData,
input: File,
subtask: SubtaskId,
testcase: TestcaseId
) -> (File, Option<Execution>)

Validate the input file editing the DAG and returing an artifact of the validator, something to keep tracks of the dependencies.

Loading content...